+++ books/books.go
@@ -1,0 +1,65 @@
+package main
+
+import (
+	"database/sql"
+	"log"
+
+	_ "github.com/lib/pq"
+)
+
+// FIXME: add publish date?
+type Book struct {
+	Title, Author string
+}
+
+type Books struct {
+	DB   *sql.DB
+	List []Book
+}
+
+func (b *Books) fetch() {
+	log.Printf("Fetching books")
+
+	rows, err := b.DB.Query(`SELECT title, author FROM book ORDER BY author ASC`)
+	if err != nil {
+		log.Fatalf("Unable to select book table:", err)
+	}
+	defer rows.Close()
+
+  b.List = []Book{}
+  var title, author string
+
+	for rows.Next() {
+		err = rows.Scan(&title, &author)
+		if err != nil {
+			log.Printf("Error: %v", err.Error())
+		}
+		b.List = append(b.List, Book{Title: title, Author: author})
+	}
+}
+
+func (b *Books) populate() {
+	log.Printf("Recreating book table")
+
+	_, err := b.DB.Query(`DROP TABLE book`)
+	if err != nil {
+		log.Fatalf("Unable to drop book table:", err)
+	}
+
+	_, err = b.DB.Query(`CREATE TABLE book
+    (id serial primary key,
+    title text NOT NULL,
+    author varchar(255) NOT NULL)`)
+	if err != nil {
+		log.Fatalf("Unable to create book table:", err)
+	}
+
+	log.Printf("Populating book table")
+
+	for _, book := range seed {
+		_, err = b.DB.Query(`INSERT INTO book (title, author) VALUES ($1,$2)`, book.Title, book.Author)
+		if err != nil {
+			log.Fatalf("Unable to populate book table:", err)
+		}
+	}
+}
