+++ books/seed.go
@@ -1,0 +1,20 @@
+package main
+
+var seed = []Book{
+	{
+		Title:  "The Hitchhiker's Guide to the Galaxy",
+		Author: "Adams, Douglas",
+	},
+	{
+		Title:  "Dirk Gently's Holistic Detective Agency",
+		Author: "Adams, Douglas",
+	},
+	{
+		Title:  "Snow Crash",
+		Author: "Stephenson, Neal",
+	},
+	{
+		Title:  "So Long, and Thanks for All the Fish",
+		Author: "Adams, Douglas",
+	},
+}
